feat: 상위 컴포넌트에서 함수 호출, bottomContent 수정
