+++ src/components/myPage/TopContent.tsx
@@ -1,45 +1,22 @@
-import axios from "axios";
-import React, { useEffect, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 
 import { UserInfo } from "../../pages/MyPage";
-import { getData } from "../../utils/lib/api";
 import { StLoginLink } from "../common/MainHeader";
 import { Button } from "../common/styled/Button";
 import { TopBanner } from ".";
 
 interface TopContentProps {
   userInfo: UserInfo;
+  isLogin: boolean;
+  handleIsLogin: () => void;
   onImageChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
 export default function TopContent(props: TopContentProps) {
   const navigate = useNavigate();
-  const { userInfo, onImageChange } = props;
-
-  const tempToken = localStorage.getItem("booktez-token");
-  const localToken = tempToken ? tempToken : "";
-
-  const [isLogin, setIsLogin] = useState<boolean>(true);
-
-  const getLogin = async (key: string, token: string) => {
-    try {
-      const { data } = await getData(key, token);
-      const status = data.status;
-
-      if (!localToken) {
-        setIsLogin(false);
-      }
-      if (!(status === 200)) {
-        setIsLogin(false);
-      }
-    } catch (err) {
-      if (axios.isAxiosError(err)) {
-        console.log("err", err.response?.data);
-      }
-    }
-  };
+  const { userInfo, isLogin, handleIsLogin, onImageChange } = props;
 
   const handleMoveLogin = (e: React.MouseEvent<HTMLButtonElement>) => {
     navigate("/login");
@@ -48,12 +25,8 @@ export default function TopContent(props: TopContentProps) {
   const handleLogout = (e: React.MouseEvent<HTMLButtonElement>) => {
     localStorage.removeItem("booktez-token");
     localStorage.removeItem("booktez-nickname");
-    setIsLogin(false);
+    handleIsLogin();
   };
-
-  useEffect(() => {
-    getLogin("/auth/check", localToken);
-  }, [isLogin]);
 
   return (
     <StWrapper>
